fix completion of render pausing other render

+++ modules/commands.py
@@ -269,8 +269,9 @@ def end(thread,self,renderdir,videoName,videopath,times,outputpath,videoQuality,
                 os.system(f'rm -rf "{thisdir}/{self.input_file}"')
                 try:
                        for i in os.listdir(f'{thisdir}'):
-                        if '.mp4' in i:
-                            os.system(f'rm -rf "{thisdir}/{i}"')
+                        if os.path.isfile(os.path.join(thisdir, i)):
+                                if '.mp4' in i:
+                                        os.system(f'rm -rf "{thisdir}/{i}"')
                             
                 except Exception as e:
                        log(str(e))
